WV-2143
- Break out GIF button and related code
